extend MAPPING_TYPE_DESCRIPTION
