+++ dynamic-mapping-ui/src/mapping/edit/edit-substitution-modal.component.ts
@@ -7,7 +7,8 @@ import {
   MappingSubstitution,
   RepairStrategy
 } from '../../shared';
-import { EditorMode, StepperConfiguration } from '../shared/stepper-model';
+import { EditorMode } from '../shared/stepper-model';
+import { StepperConfiguration } from 'src/shared/model/shared.model';
 import { definesDeviceIdentifier } from '../shared/util';
 
 @Component({
