+++ dynamic-mapping-ui/src/mapping/shared/stepper-model.ts
@@ -19,22 +19,9 @@
  * @authors Christof Strack
  */
 
-import { Direction } from '../../shared';
-
 export enum EditorMode {
   CREATE = 'CREATE',
   UPDATE = 'UPDATE',
   READ_ONLY = 'READ_ONLY',
   COPY = 'COPY'
-}
-
-export interface StepperConfiguration {
-  showEditorSource?: boolean;
-  showProcessorExtensions?: boolean;
-  editorMode: EditorMode;
-  allowNoDefinedIdentifier?: boolean;
-  allowDefiningSubstitutions?: boolean;
-  allowTestTransformation?: boolean;
-  allowTestSending?: boolean;
-  direction: Direction;
 }
