+++ dynamic-mapping-ui/src/mapping/step-three/mapping-testing.component.ts
@@ -39,7 +39,7 @@ import {
 } from '../../shared/';
 import { MappingService } from '../core/mapping.service';
 import { C8YRequest } from '../processor/processor.model';
-import { StepperConfiguration } from '../shared/stepper-model';
+import { StepperConfiguration } from 'src/shared/model/shared.model';
 import { isDisabled } from '../shared/util';
 
 @Component({
